move type to module
