+++ homeassistant/components/vicare/types.py
@@ -64,55 +64,6 @@ VICARE_TO_HA_PRESET_HEATING = {
 }
 
 
-class VentilationMode(enum.StrEnum):
-    """ViCare ventilation modes."""
-
-    PERMANENT = "permanent"  # on, speed controlled by program (levelOne-levelFour)
-    VENTILATION = "ventilation"  # activated by schedule
-    SENSOR_DRIVEN = "sensor_driven"  # activated by schedule, override by sensor
-    SENSOR_OVERRIDE = "sensor_override"  # activated by sensor
-
-    @staticmethod
-    def to_vicare_mode(mode: str | None) -> str | None:
-        """Return the mapped ViCare ventilation mode for the Home Assistant mode."""
-        if mode:
-            try:
-                ventilation_mode = VentilationMode(mode)
-            except ValueError:
-                # ignore unsupported / unmapped modes
-                return None
-            return HA_TO_VICARE_MODE_VENTILATION.get(ventilation_mode) if mode else None
-        return None
-
-    @staticmethod
-    def from_vicare_mode(vicare_mode: str | None) -> str | None:
-        """Return the mapped Home Assistant mode for the ViCare ventilation mode."""
-        for mode in VentilationMode:
-            if HA_TO_VICARE_MODE_VENTILATION.get(VentilationMode(mode)) == vicare_mode:
-                return mode
-        return None
-
-
-HA_TO_VICARE_MODE_VENTILATION = {
-    VentilationMode.PERMANENT: "permanent",
-    VentilationMode.VENTILATION: "ventilation",
-    VentilationMode.SENSOR_DRIVEN: "sensorDriven",
-    VentilationMode.SENSOR_OVERRIDE: "sensorOverride",
-}
-
-
-class VentilationProgram(enum.StrEnum):
-    """ViCare preset ventilation programs.
-
-    As listed in https://github.com/somm15/PyViCare/blob/6c5b023ca6c8bb2d38141dd1746dc1705ec84ce8/PyViCare/PyViCareVentilationDevice.py#L37
-    """
-
-    LEVEL_ONE = "levelOne"
-    LEVEL_TWO = "levelTwo"
-    LEVEL_THREE = "levelThree"
-    LEVEL_FOUR = "levelFour"
-
-
 @dataclass(frozen=True)
 class ViCareDevice:
     """Dataclass holding the device api and config."""
